Complete TODO for version change and removed unused classes(Throttler and Semaphore) (#4846)
